+++ JavaWeb-exAvaliacao/src/br/com/fiap/dao/Dao.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface Dao<T> {
 	void adicionar(T entidade);
 	List<T> listar();
-	T buscar(int id);
-	void removeById(int id);
+	T buscar(Long id);
+	void removeById(Long id);
 	void remover(T entity);
 	void update(T entity); 
 }
